first version of balance algorithm
